--- a/benches/main.rs
+++ b/benches/main.rs
@@ -19,13 +19,5 @@ fn diff_fool(c: &mut Criterion) {
     c.bench_function("diff fool", |b| b.iter(|| seqdiff::diff(&s, &t)));
 }
 
-fn bench_ratios(c: &mut Criterion) {
-    let mut group = c.benchmark_group("Ratios");
-    let s = black_box("1 1 1 1 1 1 1 1 1 1".chars().collect::<Vec<_>>());
-    let t = black_box("1 2 1 1 2 1 1 4 1 1".chars().collect::<Vec<_>>());
-    group.bench_function("normal", |b| b.iter(|| seqdiff::ratio(&s, &t)));
-    group.finish();
-}
-
-criterion_group!(benches, bench_ratios, diff_long, diff_short, diff_fool);
+criterion_group!(benches, diff_long, diff_short, diff_fool);
 criterion_main!(benches);