--- a/quickcheck
+++ b/quickcheck
@@ -1,0 +1,9 @@
+#!/usr/bin/bash
+
+while true
+do
+    cargo test $1
+    if [[ x$? != x0 ]] ; then
+        exit $?
+    fi
+done