Optimized diff algorithm. Benchmark result:

```
diff super long         time:   [21.296 ms 21.386 ms 21.493 ms]
                        change: [-97.796% -97.759% -97.724%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 14 outliers among 100 measurements (14.00%)
  4 (4.00%) high mild
  10 (10.00%) high severe

diff long               time:   [2.0008 ms 2.0088 ms 2.0192 ms]
                        change: [-95.012% -94.870% -94.726%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  4 (4.00%) high mild
  7 (7.00%) high severe

diff short              time:   [93.289 ns 94.013 ns 94.869 ns]
                        change: [-88.045% -87.973% -87.899%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 15 outliers among 100 measurements (15.00%)
  7 (7.00%) high mild
  8 (8.00%) high severe

diff fool               time:   [2.9592 ms 2.9753 ms 2.9926 ms]
                        change: [-98.397% -98.382% -98.368%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 19 outliers among 100 measurements (19.00%)
  4 (4.00%) high mild
  15 (15.00%) high severe
```